Convert repo to lerna monorepository (#16)
Convert this repo to a lerna mono-repo that will contain the data-visualizer-app, and all analytics plugins. Currently, the "app" is the only package:

the dependencies are left in the app's package.json for now until a new package is added
most of the npm scripts from root simply run the same-named script in each package. This is to make it possible to run the "same" script from root or a package.