DHIS2-4766: implement Save/SaveAs callbacks (#37)
It turned out we cannot use d2.model due to various reasons related to how models work and how the schema works on the backend.
Analytic objects are a bit special.
Use d2.api instead, POST and PUT requests for SaveAs and Save respectively. The new skipTranslations and skipSharing parameters are passed.