DHIS2-4971: Download menu component (#63)
* Add chart config from d2-charts-api to the store

This is needed for the download menu, where we need access to the
chart's SVG which is passed in the POST request to the server in order
to generate PNG/PDF for download.

* DHIS2-4971: add download menu

* Remove debug, ignore

* Change download strategy to not rely on <a download />

This opens the image/PDF in a new tab instead of triggering the download
directly. maybe it's better after all? people can look at the image/PDF
before deciding to save it in the folder they choose via the native
browser Save dialog.

* Remove revoke of url, we need it for saving the image/pdf

It's not really possible to revoke the URL as it is still needed if
one decides to save the image/PDF.

* Fix and improved tests

* Fix action type import/export for chart reducer

* Move UpdateButton from MenuBar to its own folder

* Move DownloadMenu from MenuBar to its own folder

* Put only the SVG for the chart in the store

This is an attempt to see if that was the culprit for the Redux tools to
be slow and crash sometimes.

* Fix tests for SET_CHART now that SVG string is passed

* Add strings for the download menu

* Set size for exported PNG/PDF

* Fix prettier, ignore