task/upgrade d2-ui packages and add transformIgnorePattern for jest (#105)
Upgrade to the latest d2-ui package version. With the latest version we significantly reduce the data visualizer bundle size (reduced by 15%), due to improved treeshaking.

Since d2-ui no longer comes with a commonjs build, it is necessary for jest to transform d2-ui code. Therefore the addition to the transformIgnorePatterns in the jest config.

A couple of prettier fixes (not sure where those errors came from)