DHIS2-5235: fix UI options merge when setting current
The options from UI are merged to the current object in order to avoid
loosing data that is not specifically handled (like options, axes,
type).
Deleting the keys from the options object did not work due to this
merge.
Instead, set the values for target/base line options to their defaults
(undefined), in this way the merge still works and the new values are
used.
Using undefined as value has the advantage that those keys will be
automatically removed by JSON.stringify before saving the object.