Fix "flashing" when closing snackbar
The problem is due to the animation on closing the snackbar, which takes
some time before the snackbar is not visible anymore.
Changing store.snackbar to default values causes the snackbar to show
empty and with default black styling during the animation (triggered by
setting open to false).
A solution is to split the action in 2, the first only sets open to
false, triggering the animation, the second resets store.snackbar to
default when the snackbar is not visible anymore.
The default animation duration is 195 ms, the timeout for the reset
action is set to 250.