DHIS2-5281: fix problem with LoadingMask which prevented chart to load
The error in IE11 was a DOM not found error.
My explanation is that the LoadingMask component was removed when
Highcharts replaces the content of the div with the chart.
React/IE11 does not seem to like that.
Solution is to put the LoadingMask component outside the div used by
Highcharts, LoadingMask is still controlled by the loading state prop.