Fetch chart always by id when used in dashboard
This is to avoid flashing of the chart in dashboard.
The change allows for the config object to have several keys, but still
fetch the chart by id as the config object passed from dashboard to the
plugin is a stripped down version of the AO and does not have the
necessary data for generating the chart without fetching it.