fix: avoid React warning about required proptype
"Warning: Failed prop type: The prop `items[0].name` is
marked as required in `AxisSetup`, but its value is `undefined`.
    in AxisSetup (created by WithStyles(AxisSetup))"

was sometimes popping up and it turned out sometimes there is no name
available in the metadata for the item object.
Changed the code to initialize the name to empty string when not
available instead of passing undefined.