fix: error handling in DV app/plugin
Problem was that when changing visualization type and the layout validation failed, the plugin code was still running (due to async) and generated errors due to incompatible data passed in the object.
For example, switching from a gauge/sv AO to column, the layout
validation fails because of missing dimension in category. The error was
correctly raised, but the async code in the plugin was also raising an
error which was overriding the layout error.
The solution is to run the validation before we render the Visualization
component, in App.js. In this way the error is passed directly to the
BlankCanvas component and the Visualization component is never rendered.
Also, the Visualization component does not need to check for errors and
decide if render the plugin or the BlankCanvas component.