fix: attempt to solve the various errors when switching type and on update
Layout validation is run both on viz type change and when the update
button is clicked.
This should avoid to load the plugin when there is an invalid layout
which will likely cause an error due to missing expected data/config.

Also, in the reducers we cannot always assume the layout has all the
required dimensions/items, ie. when starting a viz from scratch and
switch viz type or click the update button before selecting some
required dimension item.