feat: allow for loading of reportTable AOs
Instead of always pass "chart" in the File Menu, change to "reportTable"
when Pivot table is selected as visualization type.
This does not always currently work due to FileMenu not refreshing the
file list (it needs to be fixed in FileMenu component).
The "apiObjectName" passed around is not computed from the visualization type,
instead of just being set to "chart".