feat: WIP pivot table type (#335)
* chore: remove references to chart, use visualization instead

Now that the visualization can be also a pivot table, we should not use
chart when referring to the visualization.

* feat: add pivot table as possible visualization type

Make it the default one too (at least for now), since the AO does not
have the "type" property yet, otherwise the default (previously COLUMN)
would be used as type also for reportTable AOs.

* feat: allow for loading of reportTable AOs

Instead of always pass "chart" in the File Menu, change to "reportTable"
when Pivot table is selected as visualization type.
This does not always currently work due to FileMenu not refreshing the
file list (it needs to be fixed in FileMenu component).
The "apiObjectName" passed around is not computed from the visualization type,
instead of just being set to "chart".

* feat: load a different plugin for PIVOT_TABLE vis type

(Currently the pivot plugin is a copy of the chart one).

* chore: regenerate pot file

* fix: fix failing tests