refactor: add new options config system by viz type
Each visualization type has a set of options to show in the UI,
distributed in various tabs and sections within each tab.
Some visualization type share the exact same list of options, so
wherever possible the same config is used.
The config defines the list of tabs with their labels, and the list of
options to show in each tab.
The options can be organized in different sections; each section can
have an optional label.
Each option is a React component and is mapped to one property in the
AO.
As it was the case before, whenever an option is modified in the options
dialog, the DV redux store is immediately updated, so even if the dialog
is closed without clicking the update button, the options' state is
always saved.