refactor: use the new options config system
The visualization type is used to determine which config to use.
Processing the options config is quite simple.
First an array of tab components is generated, which are added to the
tab bar.
For each tab, an array with the components for rendering the options (possibly divided into sections)
is generated.
The currently selected tab is then activated.