refactor: change behaviour of toggable Select
If it's toggleable it means that the checkbox acts as a switch between
the default value for the option (aka disabled option) and the other
possible values listed in the Select.

The returned value used when saving the AO must then be set based on
both the Checkbox and the Select.