fix: remove custom title when Auto is selected (DHIS2-8252)
Also fixed the subtitle option to work in the same way as the title.
Before when no custom subtitle was specified, the Custom subtitle was
selected with an empty value.
Now Auto means not hidden subtitle but also not custom subtitle.
The subtitle is actually auto generated in some situations (depending on
vis type, presence of title...)