feat: fetch all legend definitions for data items
Data items can have a pre-assigned legend (done in the maintenance app) and
it should be used when the "Use pre-defined legend set per data item"
option is selected in the UI for the Legend option.

Fetch all legend sets in VisualizationPlugin and pass them via props to
chart or pivot plugin.