fix: do not pass options intended for download request DHIS2-8401
hideEmptyRows and hideEmptyColumns options are only intended for the
analytics download request, not for the normal request used to render
the table.
Passing them does not affect the response payload, but does negatively
affect caching, both client and server side, be cause the options are
passed in the query string.