fix: fix loading translations for share/translate dialogs (DHIS2-8249) (#785)
d2.init must be called after the properties files are added to i18n config, otherwise they are not requested.

d2.config.baseUrl must be set before the userSettings request.
This is to avoid an invalid request to userSettings which causes the
locale to not be set and the language properties files to not load.

Now it basically does the same thing as in v33 where we didn't use app shell.

The whole d2 will be removed eventually, we still need to pass it around
because of some d2-ui components we use that requires it. The app itself
could do without already.

The properties files are also only required by those d2-ui components.