refactor: replace MUI with ui-core for Snackbar component (TECH-323)
Because the AlertBar does not have an open prop, the component must be
used in a conditional rendering, thus it does not need to be connected
and can be simplified into a basic functional component.

App.js is where the Snackbar is used, and the logic for getting the
message for the snackbar has been moved there.

ui-core AlertBar has a default variant of type information, so for
success message the success variant need to be explicitly passed.