fix: rewrite MenuBar buttons (TECH-320) (#879)
* refactor: new MenuButton component to replace MUI buttons
* refactor: menubar style js to css module

Introduces a new custom component MenuButton to replace the old MUI buttons for Options, Download and Interpretations in the MenuBar.
https://jira.dhis2.org/browse/TECH-320

Note: As the buttons needs to match the color scheme of FileMenu (@dhis2/d2-ui-file-menu), the colors aren't fetched from ui-core just yet. There's an inline comment with a TODO about this.