feat: implement colorSet option (DHIS2-670)
The option allows to choose a color set to use in the chart.
Each color in the set is used to fill the bar/column/area/line of one
serie.
With the Mono patterns option, instead of a color, a pattern from the 10
default patterns provided by Highcharts is used.