fix: confirm dialog when navigating to/from interpretation
Since no change is actually lost when navigating in and out of an
interpretation's view, remove the confirm dialog.

Also, when answering No to the confirm dialog (meaning i don't want to
navigate to the new URL) restore the previous URL with history.goBack()
so that the app really is in the same state as before the confirm
dialog.