test: cypress tests need to confirm relevant elements exist in DOM before getting redux state (#1511)
Some tests were failing on CI but passing locally. Throttling the network connection enabled reproducing locally.

The issue was that the redux state was fetched prior to all the state changes completing. By ensuring the elements to be tested are present in the DOM before getting the redux state, we know that the redux state is current. (read about it here: https://www.cypress.io/blog/2018/11/14/testing-redux-store/)