fix: detect style changes from dashboard for re-rendering visualization (#1667)
Key features:
Visualizations in dashboard were not rerendering to fit the item size in the following situations:
* item fullscreen
* edit item resize
* change screen width (e.g., switch to/from small screen)

Description
Dashboard and Data Visualizer apps have different approaches regarding visualization size and resizing. DV vis containers have display flex and use a counter that updates when the window is resized, to trigger a re-render of the visualization. In Dashboard, the container has a set width and height, which change on any of the above resizing scenarios. It does not use a counter.

The new effect will end up being specific to dashboard, while the existing effect that responds to renderCounter will be specific to DV. A future improvement would possibly be to use the same mechanisms for both apps. But playing it safe for now, since we are in a hard freeze.