fix: explain fix headers options depending on layout selection
If one of the fix headers options is enabled and the corresponded axis
in the layout has no dimension items, show a help text under the
option explaining that there are no headers for that axis, so the option
is a noop.
The actual implementation of the check is in the pivot table engine in
analytics.