test: limit query scope to avoid long response time
This is a temporary fix for this test that started to fail on test
instances when the facility levels for the whole Sierra Leone are
requested.

Also, wait for the org unit tree to be loaded before trying to select
anything else than the root level, this seemed to fail consistently from
my local testing.