fix: plugin flashes when interacting with Interpretations modal (#3017)
Fixes DHIS2-15570
Fixes DHIS2-17137

Prevent re-requests for analytics, and resulting re-renders that were giving a flashing effect.

* InterpretationModal: memoize the callback to avoid a re-render due to a "new" function
being passed with each render of InterpretationModal.
The effect of this change is minor, probably just 1 rerender

* VisualizationPlugin: ensure that the size of the plugin is set when it is coming
from the modal so that when changing the width of the viewport,
the visualization adjusts its size accordingly. (Previously this "worked" but only
because of all the re-fetch/re-renders)

* VisualizationPluginWrapper: need to listen to the individual properties,
rather than the whole props object since that was triggering lots of
unnecessary re-fetch/renders.

* set the max-height to respect viewport so that vertical resize causes
the chart to adjust accordingly