fix: reset error when some props change (DHIS2-18576)
This fixes the issue of DV plugins getting "stuck" in showing an error
even after the cause is removed.
If a dashboard filter is applied and causes a DV visualization to not
have data, an error is shown in the plugin.
When the filter was removed, the error was still shown.