fix: fetch and extract option set id
This information is not returned as part of columns/rows/filters but it
can be extracted from dataDimensionItems.
The optionSetId is also added to the metadata when fetching the
dataItems in the Data modal.
The rest of the code relies on the presence of optionSetId in metadata
in order to enable the option view mode.