test: ensure the data item click does not happen on optional buttons
Default behaviour for the click/dblclick is to position the cursor in
the middle of the element.
Depending on the data item, the option view mode button might be
positioned right in the middle of the element, causing the
click/dblclick to trigger the option view mode.
The left side of the data item is a good solution since it does not have
any button.