feat: implement option set selector in Data dimension modal (DHIS2-17872) (#3344)
Implements https://dhis2.atlassian.net/browse/DHIS2-17872
Requires https://github.com/dhis2/analytics/pull/1746

The "option view mode" is enabled via the list icon added to the data items with an assigned optionSetId.
When this mode is enabled, a separate Transfer component is shown in place of the "primary" one, which is just hidden but still present in the DOM.
This allows for all selections, highlights, filtering, etc... to persist after the "option view mode" is dismissed.
Both Transfer components are affecting the same selected items list and therefore changes made in one are automatically reflected in the other.

The "option view mode" Transfer component is using filters for fetching program data element options or program attribute options depending on the clicked data item.
The source (left) part of the Transfer is only listing the options for a specific data item/type.
The selected (right) part of the Transfer has the same list as the "primary" Transfer component.