fix: restore navigation between visualizations using the browser address bar [DHIS2-19387] (#3375)
Implements DHIS2-19387

Manual URL changes trigger location updates with the key default,
which then results in the history skipping its handler because it's the
same as the key as the previous updates.

It's still important to check they key to avoid triggering the handler multiple
times for the same location update, but the pathname and search can be
checked to see if the rest of the handler should run.

Co-authored-by: Jen Jones Arnesen <jennifer@dhis2.org>